+++ src/main/controllers/ChatRoomScreenController.java
@@ -1,6 +1,7 @@
 package main.controllers;
 
 import main.presenters.ChatRoomScreen;
+import main.screencontrollers.MessageScreenController;
 import main.screencontrollers.ScreenController;
 
 import java.time.LocalDateTime;
